+++ components/Header/src/index.tsx
@@ -1,9 +1,9 @@
 import './index.scss';
 
-export * from './HeaderLogic';
 export * from './Header';
+export * from './HeaderBase';
 export * from './HeaderAction';
 export * from './HeaderActions';
 export * from './HeaderContent';
-export * from './HeaderLogoContainer';
 export * from './HeaderMobileActions';
+export * from './HeaderMobileToggle';