+++ components/LanguageSwitcher/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @gemeente-denhaag/language-switcher
 
+## 3.2.3
+
+### Patch Changes
+
+- 068c53d: Internal tsconfig changes
+- Updated dependencies [12c575c]
+- Updated dependencies [068c53d]
+  - @gemeente-denhaag/icons@4.3.0
+  - @gemeente-denhaag/link@4.2.2
+  - @gemeente-denhaag/link-button@3.2.2
+
 ## 3.2.2
 
 ### Patch Changes