+++ components/SkipLink/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @gemeente-denhaag/skip-link
 
+## 1.2.2
+
+### Patch Changes
+
+- 068c53d: Internal tsconfig changes
+
 ## 1.2.1
 
 ### Patch Changes