+++ packages/keycloak/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gemeente-denhaag/keycloak
 
+## 3.2.5
+
+### Patch Changes
+
+- Updated dependencies [12c575c]
+- Updated dependencies [068c53d]
+  - @gemeente-denhaag/design-tokens@1.5.0
+  - @gemeente-denhaag/components-css@4.1.1
+
 ## 3.2.4
 
 ### Patch Changes